+++ pkg/ddl/BUILD.bazel
@@ -93,7 +93,6 @@ go_library(
         "//pkg/disttask/framework/taskexecutor/execute",
         "//pkg/disttask/operator",
         "//pkg/domain/infosync",
-        "//pkg/domain/resourcegroup",
         "//pkg/errctx",
         "//pkg/expression",
         "//pkg/expression/context",
@@ -119,6 +118,7 @@ go_library(
         "//pkg/parser/terror",
         "//pkg/parser/types",
         "//pkg/privilege",
+        "//pkg/resourcegroup",
         "//pkg/resourcemanager/pool/workerpool",
         "//pkg/resourcemanager/util",
         "//pkg/sessionctx",
