+++ pkg/distsql/context/BUILD.bazel
@@ -6,10 +6,10 @@ go_library(
     importpath = "github.com/pingcap/tidb/pkg/distsql/context",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/domain/resourcegroup",
         "//pkg/errctx",
         "//pkg/kv",
         "//pkg/parser/mysql",
+        "//pkg/resourcegroup",
         "//pkg/util/context",
         "//pkg/util/execdetails",
         "//pkg/util/memory",
