+++ pkg/domain/BUILD.bazel
@@ -37,7 +37,6 @@ go_library(
         "//pkg/domain/globalconfigsync",
         "//pkg/domain/infosync",
         "//pkg/domain/metrics",
-        "//pkg/domain/resourcegroup",
         "//pkg/errno",
         "//pkg/infoschema",
         "//pkg/infoschema/metrics",
@@ -55,6 +54,7 @@ go_library(
         "//pkg/parser/terror",
         "//pkg/planner/core/metrics",
         "//pkg/privilege/privileges",
+        "//pkg/resourcegroup/runaway",
         "//pkg/sessionctx",
         "//pkg/sessionctx/sessionstates",
         "//pkg/sessionctx/sysproctrack",
@@ -66,8 +66,6 @@ go_library(
         "//pkg/statistics/handle/logutil",
         "//pkg/statistics/handle/util",
         "//pkg/store/helper",
-        "//pkg/ttl/cache",
-        "//pkg/ttl/sqlbuilder",
         "//pkg/ttl/ttlworker",
         "//pkg/types",
         "//pkg/util",
@@ -98,7 +96,6 @@ go_library(
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_kvproto//pkg/metapb",
         "@com_github_pingcap_kvproto//pkg/pdpb",
-        "@com_github_pingcap_kvproto//pkg/resource_manager",
         "@com_github_pingcap_log//:log",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//oracle",
