+++ pkg/domain/infosync/BUILD.bazel
@@ -20,13 +20,13 @@ go_library(
         "//pkg/ddl/label",
         "//pkg/ddl/placement",
         "//pkg/ddl/util",
-        "//pkg/domain/resourcegroup",
         "//pkg/errno",
         "//pkg/kv",
         "//pkg/metrics",
         "//pkg/parser/model",
         "//pkg/parser/mysql",
         "//pkg/parser/terror",
+        "//pkg/resourcegroup",
         "//pkg/session/cursor",
         "//pkg/sessionctx/binloginfo",
         "//pkg/sessionctx/variable",
