+++ pkg/executor/BUILD.bazel
@@ -103,7 +103,6 @@ go_library(
         "//pkg/disttask/importinto",
         "//pkg/domain",
         "//pkg/domain/infosync",
-        "//pkg/domain/resourcegroup",
         "//pkg/errctx",
         "//pkg/errno",
         "//pkg/executor/aggfuncs",
@@ -161,6 +160,8 @@ go_library(
         "//pkg/plugin",
         "//pkg/privilege",
         "//pkg/privilege/privileges",
+        "//pkg/resourcegroup",
+        "//pkg/resourcegroup/runaway",
         "//pkg/resourcemanager/pool/workerpool",
         "//pkg/resourcemanager/util",
         "//pkg/session/txninfo",
