+++ pkg/executor/internal/querywatch/BUILD.bazel
@@ -7,13 +7,14 @@ go_library(
     visibility = ["//pkg/executor:__subpackages__"],
     deps = [
         "//pkg/domain",
-        "//pkg/domain/resourcegroup",
         "//pkg/executor/internal/exec",
         "//pkg/infoschema",
         "//pkg/parser",
         "//pkg/parser/ast",
         "//pkg/parser/model",
         "//pkg/planner/util",
+        "//pkg/resourcegroup",
+        "//pkg/resourcegroup/runaway",
         "//pkg/sessionctx",
         "//pkg/util/chunk",
         "@com_github_pingcap_errors//:errors",
