+++ pkg/resourcegroup/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "resourcegroup",
+    srcs = ["checker.go"],
+    importpath = "github.com/pingcap/tidb/pkg/resourcegroup",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@com_github_pingcap_kvproto//pkg/resource_manager",
+        "@com_github_tikv_client_go_v2//tikvrpc",
+    ],
+)
