+++ pkg/resourcegroup/runaway/BUILD.bazel
@@ -1,16 +1,35 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "resourcegroup",
-    srcs = ["runaway.go"],
-    importpath = "github.com/pingcap/tidb/pkg/domain/resourcegroup",
+    name = "runaway",
+    srcs = [
+        "checker.go",
+        "manager.go",
+        "record.go",
+        "syncer.go",
+    ],
+    importpath = "github.com/pingcap/tidb/pkg/resourcegroup/runaway",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/ddl",
+        "//pkg/infoschema",
+        "//pkg/kv",
         "//pkg/metrics",
+        "//pkg/parser/model",
+        "//pkg/parser/terror",
+        "//pkg/sessionctx",
+        "//pkg/ttl/cache",
+        "//pkg/ttl/sqlbuilder",
+        "//pkg/types",
+        "//pkg/util",
+        "//pkg/util/chunk",
         "//pkg/util/dbterror/exeerrors",
         "//pkg/util/generic",
         "//pkg/util/logutil",
+        "//pkg/util/sqlexec",
         "@com_github_jellydator_ttlcache_v3//:ttlcache",
+        "@com_github_pingcap_errors//:errors",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_kvproto//pkg/resource_manager",
         "@com_github_prometheus_client_golang//prometheus",
         "@com_github_tikv_client_go_v2//tikv",
