+++ pkg/server/BUILD.bazel
@@ -24,7 +24,6 @@ go_library(
         "//pkg/config",
         "//pkg/domain",
         "//pkg/domain/infosync",
-        "//pkg/domain/resourcegroup",
         "//pkg/errno",
         "//pkg/executor",
         "//pkg/executor/mppcoordmanager",
@@ -48,6 +47,7 @@ go_library(
         "//pkg/privilege/conn",
         "//pkg/privilege/privileges",
         "//pkg/privilege/privileges/ldap",
+        "//pkg/resourcegroup",
         "//pkg/server/err",
         "//pkg/server/handler",
         "//pkg/server/handler/extractorhandler",
