+++ pkg/server/metrics/BUILD.bazel
@@ -6,9 +6,9 @@ go_library(
     importpath = "github.com/pingcap/tidb/pkg/server/metrics",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/domain/resourcegroup",
         "//pkg/metrics",
         "//pkg/parser/mysql",
+        "//pkg/resourcegroup",
         "@com_github_prometheus_client_golang//prometheus",
     ],
 )
