+++ pkg/store/copr/BUILD.bazel
@@ -19,11 +19,11 @@ go_library(
         "//pkg/config",
         "//pkg/ddl/placement",
         "//pkg/domain/infosync",
-        "//pkg/domain/resourcegroup",
         "//pkg/errno",
         "//pkg/kv",
         "//pkg/metrics",
         "//pkg/parser/terror",
+        "//pkg/resourcegroup",
         "//pkg/sessionctx/variable",
         "//pkg/store/copr/metrics",
         "//pkg/store/driver/backoff",
